requirements.txt
# Core dependencies
fastapi==0.95.0
streamlit==1.22.0
pymupdf==1.21.1
python-docx==0.8.11
spacy==3.5.0
transformers==4.26.1
torch==2.0.0
scikit-learn==1.2.2
numpy==1.24.2
pandas==2.0.0

# Optional dependencies
nltk==3.7